convert sample to lit test
