+++ frontend/test/pytest/test_measurement_primitives.py
@@ -36,148 +36,7 @@ from catalyst.jax_tracer import lower_jaxpr_to_mlir
 from catalyst.tracing.contexts import EvaluationContext
 
 
-@pytest.mark.xfail(reason="[WIP] Convert to lit test")
-def test_sample_static():
-    """Test that the sample primitive can be captured into jaxpr."""
-
-    def f():
-        obs = compbasis_p.bind()
-        return sample_p.bind(obs, shots=5, num_qubits=0)
-
-    jaxpr = jax.make_jaxpr(f)().jaxpr
-    mlir = lower_jaxpr_to_mlir(jax.make_jaxpr(f)(), "foo")[0]
-    print(jaxpr)
-    print(mlir)
-    assert (
-        jaxpr
-        == """
-{ lambda ; . let
-    a:AbstractObs(num_qubits=0,primitive=compbasis) = compbasis
-    b:f64[5,0] = sample[num_qubits=0 shots=5] a
-  in (b,) }
-"""
-    )
-
-    assert (
-        mlir
-        == """
-module @foo {
-  func.func public @jit_foo() -> tensor<5x0xf64> {
-    %0 = "quantum.compbasis"() : () -> !quantum.obs
-    %1 = "quantum.sample"(%0) : (!quantum.obs) -> tensor<5x0xf64>
-    return %1 : tensor<5x0xf64>
-  }
-}
-"""
-    )
-
-    # assert jaxpr.eqns[1].primitive == sample_p
-    # assert jaxpr.eqns[1].params == {"shape": (5, 0), "shots": 5}
-    # assert jaxpr.eqns[1].outvars[0].aval.shape == (5, 0)
-
-
-@pytest.mark.xfail(reason="[WIP] Convert to lit test")
-def test_sample_dynamic_shape():
-    """Test that the sample primitive with dynamic shape can be captured into jaxpr."""
-
-    @catalyst.qjit
-    @qml.qnode(
-        qml.device("null.qubit", wires=1)
-    )  # SampleOp is only legal if there is a device in the same scope
-    def f(shots: int):
-        obs = compbasis_p.bind()
-        x = shots + 1
-        sample = sample_p.bind(obs, x, num_qubits=0)
-        return sample + jax.numpy.zeros((x, 0))
-
-    print(f.jaxpr)
-    print(f.mlir)
-    # breakpoint()
-
-    # jaxpr_cat = catalyst.jax_extras.tracing.make_jaxpr2(f)(5)
-    # breakpoint()
-    # jaxpr = jax.make_jaxpr(f)(5).jaxpr
-    # breakpoint()
-    # mlir = lower_jaxpr_to_mlir(jax.make_jaxpr(f)(5), "foo")[0]
-    # print(jaxpr)
-    # print(mlir)
-    assert (
-        jaxpr
-        == """
-{ lambda ; a:i64[]. let
-    transform_named_sequence
-    b:i64[] c:f64[b,0] = quantum_kernel[
-      call_jaxpr={ lambda ; d:i64[]. let
-          e:AbstractObs(num_qubits=0,primitive=compbasis) = compbasis
-          f:i64[] = add d 1
-          g:f64[f,0] = sample[num_qubits=0] e f
-          h:f64[f,0] = broadcast_in_dim[broadcast_dimensions=() shape=(None, 0)] 0.0
-            f
-          i:f64[f,0] = add g h
-          qdevice[
-            rtd_kwargs={}
-            rtd_lib=/home/paul.wang/catalyst_new/catalyst/frontend/catalyst/utils/../../../runtime/build/lib/librtd_null_qubit.so
-            rtd_name=NullQubit
-          ] 0
-          j:AbstractQreg() = qalloc 1
-          qdealloc j
-        in (f, i) }
-      qnode=<QNode: device='<null.qubit device (wires=1) at 0x74a42d0abeb0>', interface='auto', diff_method='best'>
-    ] a
-  in (b, c) }
-"""
-    )
-
-    assert (
-        mlir
-        == """
-module @f {
-  func.func public @jit_f(%arg0: tensor<i64>) -> (tensor<i64>, tensor<?x0xf64>) attributes {llvm.emit_c_interface} {
-    %0:2 = catalyst.launch_kernel @module_f::@f(%arg0) : (tensor<i64>) -> (tensor<i64>, tensor<?x0xf64>)
-    return %0#0, %0#1 : tensor<i64>, tensor<?x0xf64>
-  }
-  module attributes {transform.with_named_sequence} {
-    transform.named_sequence @__transform_main(%arg0: !transform.op<"builtin.module">) {
-      transform.yield
-    }
-  }
-  module @module_f {
-    func.func public @f(%arg0: tensor<i64>) -> (tensor<i64>, tensor<?x0xf64>) attributes {diff_method = "parameter-shift", llvm.linkage = #llvm.linkage<internal>, qnode} {
-      %c0_i64 = arith.constant 0 : i64
-      %c = stablehlo.constant dense<0> : tensor<1xi32>
-      %cst = stablehlo.constant dense<0.000000e+00> : tensor<f64>
-      %c_0 = stablehlo.constant dense<1> : tensor<i64>
-      %0 = quantum.compbasis  : !quantum.obs
-      %1 = stablehlo.add %arg0, %c_0 : tensor<i64>
-      %2 = quantum.sample %0 : tensor<?x0xf64>
-      %3 = stablehlo.convert %1 : (tensor<i64>) -> tensor<i32>
-      %4 = stablehlo.reshape %3 : (tensor<i32>) -> tensor<1xi32>
-      %5 = stablehlo.concatenate %4, %c, dim = 0 : (tensor<1xi32>, tensor<1xi32>) -> tensor<2xi32>
-      %6 = stablehlo.dynamic_broadcast_in_dim %cst, %5, dims = [] : (tensor<f64>, tensor<2xi32>) -> tensor<?x0xf64>
-      %7 = stablehlo.add %2, %6 : tensor<?x0xf64>
-      quantum.device shots(%c0_i64) ["/home/paul.wang/catalyst_new/catalyst/frontend/catalyst/utils/../../../runtime/build/lib/librtd_null_qubit.so", "NullQubit", "{}"]
-      quantum.device_release
-      return %1, %7 : tensor<i64>, tensor<?x0xf64>
-    }
-  }
-  func.func @setup() {
-    quantum.init
-    return
-  }
-  func.func @teardown() {
-    quantum.finalize
-    return
-  }
-}
-"""
-    )
-
-    # assert jaxpr.eqns[1].primitive == sample_p
-    # assert jaxpr.eqns[1].params == {"shape": (5, 0), "shots": 5}
-    # assert jaxpr.eqns[1].outvars[0].aval.shape == (5, 0)
-
-
-def test_new_sampleop_still_good_with_backend():
+def test_dynamic_sample_backend_functionality():
     """Test that a `sample` program with dynamic shots can be executed correctly."""
 
     @catalyst.qjit(keep_intermediate=True)
